% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyff.R
\name{activate}
\alias{activate}
\alias{active}
\alias{activate}
\alias{activate.tidyff}
\title{activate}
\usage{
active(x, value)

\method{activate}{tidyff}(.data, what)
}
\arguments{
\item{x}{object}

\item{value}{activation name}

\item{.data}{object with sub-element}

\item{what}{sub-element to activate}
}
\description{
activate
}
