% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-ini.R
\name{dim_order}
\alias{dim_order}
\alias{dim_order.BasicRaster}
\alias{dim_order.character}
\title{dimension order}
\usage{
dim_order(x, ...)

\method{dim_order}{BasicRaster}(x, ...)

\method{dim_order}{character}(x, ...)
}
\arguments{
\item{x}{filename or raster object or ff object}

\item{...}{ignored}
}
\value{
integer vector of dimension positions (1-based)
}
\description{
The dimension order as used by ff and raster native grid, see `raster::writeRaster`. 
The raster package uses the named conventions "BSQ", "BIL" and
"BIP" which are domain-specialized ways to record the dimension 
order of (virtual) 3D arrays. They stand for "band sequential"
"band interleaved" and ... correspond to c(1, 2, 3) etc. etc.
}
\details{
This function always returns an integer vector indicating the order
of the dims of a `length(dim_order(x))` array oriented relative
to the R array convention.
}
\examples{
r <- raster::raster(volcano)
dim_order(r)
w <- raster::writeRaster(r, raster::rasterTmpFile(), bandorder = "BIP")
dim_order(w)
dim_order(raster::filename(w))
}
