% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffraster.R
\name{ff_object}
\alias{ff_object}
\alias{ff_object.BasicRaster}
\alias{ff_object.character}
\title{Raster and 'ff' file-backed arrays}
\usage{
ff_object(x, readonly = TRUE, filename = NULL, ...)

\method{ff_object}{BasicRaster}(x, readonly = TRUE, filename = NULL,
  ...)

\method{ff_object}{character}(x, readonly = TRUE, filename = NULL, ...)
}
\arguments{
\item{x}{raster or raster-able file}

\item{readonly}{open in read-only mode (TRUE by default)}

\item{filename}{path to file to create}

\item{...}{arguments to methods}
}
\value{
`ff` object
}
\description{
Create an file-backed 'ff' object from a raster, or a raster-enabled file.
}
\details{
When the object is created from a filename, `raster::brick` is used. The 'dim'ension
for raster and for ff in this context always keeps degenerate singletons. Please
get in touch if this causes you problems.
}
\examples{
f <- system.file("extdata", "raster", "sst.grd", package  = "ffraster")
ff_object(raster::brick(f))
if (interactive()) {
  arr <- ff_object(f, filename = "afile.grd")
 }
}
