% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_raster_create.R
\name{ff_raster}
\alias{ff_raster}
\title{Create an ffraster}
\usage{
ff_raster(x, nlayers = NULL, filename = NULL, ..., setZ = NULL)
}
\arguments{
\item{x}{a raster object, used as the template for the data to come}

\item{nlayers}{the number of layers to instantiate}

\item{filename}{the .grd filename}

\item{...}{arguments passed to .writeGRD}

\item{setZ}{optional z values, must be of length 'nlayers'}
}
\description{
Use a raster template to set up an ff array for filling.
}
\details{
Note that no data is transferred, this is a set up function to give an ff array to be populated.

There is currently no  control over the bandorder, set to band interleaved (BIL) currently.
}
\examples{
\dontrun{
n <- 5
files <- raadtools::sstfiles()[1:n, ]

library(raster)
fun <- function(date) raadtools::readsst(date, xylim = extent(120, 160, -50, -30), inputfiles = files)
arr <- ff_raster(fun(files$date[1]), nlayers = n, filename = "afile.grd", overwrite = TRUE)
for (i in seq_along(files$date)) {
arr[,,i] <-   raster::values(t(fun(files$date[i])))
}
# plot(brick("afile.grd"))
}
}
