% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-ini.R
\name{ff_type}
\alias{ff_type}
\alias{ff_type}
\alias{ff_type.character}
\alias{ff_type}
\alias{ff_type.BasicRaster}
\alias{ff_type}
\alias{ff_type.raster_ini}
\title{ff type from raster type}
\usage{
ff_type(x, ...)

\method{ff_type}{character}(x, ...)

\method{ff_type}{BasicRaster}(x, ...)

\method{ff_type}{raster_ini}(x, ...)
}
\arguments{
\item{x}{filename, raster object, or ini object}

\item{...}{ignored}
}
\value{
type of ff, see `ff::vmode`
}
\description{
ff type from raster type
}
\examples{
ff_type("LOG1S")
}
