% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyff.R
\name{hyper_filter}
\alias{hyper_filter}
\alias{hyper_filter.tidyff}
\title{hyper filter}
\usage{
\method{hyper_filter}{tidyff}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{arguments as per `tidync::hyper_filter`}
}
\description{
hyper filter
}
