% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-ini.R
\name{ini_file}
\alias{ini_file}
\alias{ini_file.character}
\alias{ini_file.BasicRaster}
\title{raster ini file}
\usage{
ini_file(x)

\method{ini_file}{character}(x)

\method{ini_file}{BasicRaster}(x)
}
\arguments{
\item{x}{file name or raster}
}
\description{
Read the configuration file in raw list form, you can 
use either the filename (.grd) or a raster object.
}
\examples{
f <- system.file("extdata", "raster", "sst.grd", package  = "ffraster")
ini_file(f)
}
